% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.variable.bulk.R
\name{add.variable.bulk}
\alias{add.variable.bulk}
\title{Append empty columns to an existing data frame.}
\usage{
add.variable.bulk(df, l)
}
\arguments{
\item{df}{A data frame}

\item{l}{A list of the empty columns to be appended}
}
\value{
A new data frame with newly generated empty columns
}
\description{
Append empty columns to an existing data frame.
}
\examples{
\dontrun{
new.df <- add.variable.bulk(df,c("cell.line","cell.replicate","date.performed"))
}
}
