% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.variable.R
\name{add.variable.one}
\alias{add.variable.one}
\title{Append one empty columns to an existing data frame.}
\usage{
add.variable.one(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A new data frame with newly generated empty columns
}
\description{
The function will prompt for the name of the new column
}
\examples{
\dontrun{
new.df <- add.variable(df)
}
}
