% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.qPCR.R
\name{analyze.qPCR}
\alias{analyze.qPCR}
\title{analyze.qPCR()}
\usage{
analyze.qPCR()
}
\value{
A list with the plate map at each step of the analysis process.
}
\description{
Will analyze a raw RT-qPCR .csv output using the standard 2^(−ddCt) Method, a classict calculation of the fold change.
There is not input needed but once started the function will prompt for information about the plate map used.
The .csv file used need to have column names on the fist row and the associated raw data in the subsequent rows.
Designed for 96 wells plates, will work for other plate size depending on raw file format.
}
\examples{
\dontrun{
result <- analyze.qPCR()
}
}
