% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.data.R
\name{clean.data}
\alias{clean.data}
\title{Remove rows in a data frame.}
\usage{
clean.data(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A new data frame without the dropped rows
}
\description{
Will remove rows in a data frame when either one variable matches a value in said row or when said row matches two separate values.
Easy method to remove replicate associated with a specif Primer and/or Sample # in a master file.
The function only asks for a data frame, the function will then prompt for additional questions.
}
\examples{
\dontrun{
new.df <- clean.data(df)
}
}
