% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.3sub.average.R
\name{df.3sub.unique.av}
\alias{df.3sub.unique.av}
\title{Average technical replicates into biological replicates.}
\usage{
df.3sub.unique.av(
  df,
  subset1.var.name,
  subset2.var.name,
  subset3.var.name,
  data,
  data.add
)
}
\arguments{
\item{df}{A data frame}

\item{subset1.var.name}{Name of the first subsetting variable (e.g. Primer). Has to match a column name in the data frame.}

\item{subset2.var.name}{Name of the second subsetting variable (e.g. Condition). Has to match a column name in the data frame.}

\item{subset3.var.name}{Name of the third subsetting variable (e.g. Sample). Has to match a column name in the data frame.}

\item{data}{Name of the column containing the values to average. Has to match a column name in the data frame.}

\item{data.add}{Additional variable you wish to carry over to the new data frame. Has to match a column name in the data frame.}
}
\value{
A new data frame with average of technical replicates.
}
\description{
Will average technical replicates into biological replicates.
}
\examples{
\dontrun{
new.df <- df.3sub.unique.av(df,"Primer","inflammatory.factor","Sample","exp.data","cell.line")
}
}
