% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.t.test.R
\name{df.t.test}
\alias{df.t.test}
\title{Perform a serie of t-test based on two subsetting variable.}
\usage{
df.t.test(df, subset1.var.name, subset2.var.name, t.test.var1, t.test.var2)
}
\arguments{
\item{df}{A data frame}

\item{subset1.var.name}{Name of the first subsetting variable (e.g. Primer). Has to match a column name in the data frame.}

\item{subset2.var.name}{Name of the second subsetting variable (e.g. Condition). Has to match a column name in the data frame.}

\item{t.test.var1}{Name of the column containing the data. Has to match a column name in the data frame.}

\item{t.test.var2}{Name of the grouping variable for the t-test. Has to match a column name in the data frame.}
}
\value{
A new data frame with a list of p-value for all grouping combinations.
}
\description{
A time saver when working with  large number of Primer and experimental condition.
}
\examples{
\dontrun{
new.df <- df.t.test(df,"Primer","inflammatory.factor","exp.data","cell.line")
}
}
