% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.variable.R
\name{fill.variable}
\alias{fill.variable}
\title{An intuitive way to add missing information to analysed values.}
\usage{
fill.variable(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A new data frame with the newly filled-in information.
}
\description{
Will fill a variable in a data frame when either one secondary variable matches a value in said row or when said row matches two separate secondary values.
The function only asks for a data frame, the function will then prompt for additional questions.
}
\examples{
\dontrun{
new.df <- fill.variable(df)
}
}
