% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.to.df.R
\name{list.to.df}
\alias{list.to.df}
\title{Convert analyse.qPCR() outputs into an easy to read data frame.}
\usage{
list.to.df(df)
}
\arguments{
\item{df}{A list generated by analyse.qPCR()}
}
\value{
A new data frame with the newly filled-in information.
}
\description{
Will convert the list generated by analyse.qPCR() into an easy to read data frame.
The output of analyse.qPCR() HAS TO go through this finction before being used with any other function part of the WesCell package.
}
\examples{
\dontrun{
new.df <- list.to.df(list)
}
}
